third_party/blink/web_tests/external/wpt/editing/run/formatblock_1-1000-expected.txt

This is a testharness.js-based test.
Found 20 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "{<p><p> <p>foo</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div> foo</div>" but got "<p></p><p> </p><div>foo</div>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "{<p><p> <p>foo</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div> foo</div>" but got "<p></p><p> </p><div>foo</div>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "div" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "div" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "div" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "div" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><div>foo</div></td><td><div>bar</div></td><td><div>baz</div></td></tr></tbody></table>" but got "<div><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></div>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><div>foo</div></td><td><div>bar</div></td><td><div>baz</div></td></tr></tbody></table>" but got "<div><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></div>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><div>foo</div></td><td><div>bar</div></td><td><div>baz</div></td></tr></tbody></table>" but got "<div><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></div>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><div>foo</div></td><td><div>bar</div></td><td><div>baz</div></td></tr></tbody></table>" but got "<div><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></div>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<dl><dt>[foo]<dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><dl><dd>bar</dd></dl>" but got "<dl><div>foo</div><dd>bar</dd></dl>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<dl><dt>[foo]<dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><dl><dd>bar</dd></dl>" but got "<dl><div>foo</div><dd>bar</dd></dl>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<dl><dt>foo<dd>[bar]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt></dl><div>bar</div>" but got "<dl><dt>foo</dt><div>bar</div></dl>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<dl><dt>foo<dd>[bar]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt></dl><div>bar</div>" but got "<dl><dt>foo</dt><div>bar</div></dl>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<dl><dt>[foo<dd>bar]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div>bar</div>" but got "<dl><div>foo<br>bar</div></dl>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<dl><dt>[foo<dd>bar]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div>bar</div>" but got "<dl><div>foo<br>bar</div></dl>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<ol><li>[foobar]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><div>foobar</div></li></ol>" but got "<div><ol><li>foobar</li></ol></div>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<ol><li>[foobar]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><div>foobar</div></li></ol>" but got "<div><ol><li>foobar</li></ol></div>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<ul><li>[foobar]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li><div>foobar</div></li></ul>" but got "<div><ul><li>foobar</li></ul></div>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<ul><li>[foobar]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li><div>foobar</div></li></ul>" but got "<div><ul><li>foobar</li></ul></div>"
Harness: the test ran to completion.
